user can now change their db-settings

+++ app/src/main/java/fragment/user/SignupFragment.kt
@@ -124,7 +124,7 @@ class SignupFragment : Fragment() {
                 name = name,
                 email = email,
                 address = address,
-                phoneNumber = phoneNumber.toString().toInt(),
+                phoneNumber = phoneNumber.toString(),
                 uID = user?.uid
             )
